project/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
chromadb==0.4.24
numpy==1.26.4
sentence-transformers==2.7.0
pypdf2==3.0.1
python-docx==1.1.2
openai==1.30.1
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.7
